requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
python-dotenv==1.0.0
reportlab==4.0.8
Pillow==10.1.0
flask-swagger-ui==4.11.1
flasgger==0.9.7.1
stripe==7.5.0
gunicorn==21.2.0
